Added kluge iframe and fixed small mistake with bat collision. Bat now has 3hp for now and will die after 3 hits

+++ project_code/src/hitbox_system/systems.rs
@@ -5,12 +5,20 @@ use bevy::sprite::MaterialMesh2dBundle;
 
 // System to check collisions between hitboxes and hurtboxes
 pub fn check_hitbox_hurtbox_collisions(
+    time: Res<Time>,
     hitbox_query: Query<(Entity, &Transform, &Hitbox)>,
     mut hurtbox_query: Query<(Entity, &Transform, &mut Hurtbox)>,
     mut commands: Commands,
 ) {
     // // Iterate through all entities with hurtboxes
     for (hurtbox_entity, hurtbox_transform, mut hurtbox) in hurtbox_query.iter_mut() {
+        if !hurtbox.iframe.finished() {
+            hurtbox.iframe.tick(time.delta());
+            //println!("Timer left: {}", hurtbox.iframe.remaining_secs());
+
+            continue;
+        }
+
         let hurtbox_pos = hurtbox_transform.translation.truncate() + hurtbox.offset;
         let hurtbox_min = hurtbox_pos - hurtbox.size / 2.0;
         let hurtbox_max = hurtbox_pos + hurtbox.size / 2.0;
@@ -28,6 +36,7 @@ pub fn check_hitbox_hurtbox_collisions(
                 {
                     hurtbox.colliding = true;
                     println!("{} collided with {}", hurtbox.entity, hitbox.entity);
+                    hurtbox.iframe = Timer::from_seconds(1., TimerMode::Once);
                     break;
                 }
             }
@@ -125,6 +134,7 @@ pub fn create_hurtbox(
         offset,
         entity: entity_type,
         colliding: false,
+        iframe: Timer::from_seconds(1., TimerMode::Once),
     });
 }
